requirements.txt
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
future==0.17.1
numpy==1.17.1
opencv-python==4.1.0.25
Pillow==6.1.0
